changes in MG5-NLO gridpack workflow for scale/PDF reweighting without using systematics.py
The default workflow in the MG5 gridpack is to first generate the events in LHE format (1st step),
and then add to the LHE file the weights corresponding to scale and PDF variations (2nd step).
This 2nd step is done with the executable ./bin/aMCatNLO using the functionalities
defined in systematics.py to calculate the relevant weights on-the-fly.
Since systematics.py does not take into acc

The xsec calculation in systematics.py is not exactly the same as
in the fortran routine used by MG5 when generating the events (see SubProcesses/reweight_xsec_events.f);
in particular, the python module is missing the factor rwgt_mur_dep_fac,
which is present in the fortran code, although it equals unity by default.

If one does use the function rwgt_mur_dep_fac to calculate
a scale-dependent factor in the xsec (e.g. running of a Yukawa coupling),
the varied-scale/pdf weights cannot be recomputed a posteriori with systematics.py
(as that will not recalculate rwgt_mur_dep_fac).

One solution is to modify the workflow in the gridpack and
produce both events and varied-scale weights in one go (using ./generate_events).

The changes in this commit are the minimal ones to obtain this;
the manual addition of flags to disable the reweighting in the 1st step are commented
(they need to be set correctly in the run_card provided by the user),
and the execution of the 2nd step using systematics.py is disabled.